Update maestro.py
error is due to a common issue when working with threads and Tkinter. The problem is that you're trying to update the Tkinter GUI from a non-main thread, which is not allowed in Tkinter.
Here's how we can fix this:

Instead of directly updating the GUI from the MarketDataListener thread, we'll use a queue to communicate between threads.

We'll modify the schedule_update_market_data_status method to use the queue instead of directly calling Tkinter methods.
We'll add a method to periodically check this queue and update the GUI from the main thread.